Use long instead of byte

+++ musig2/src/main/java/com/chainx/musig2bitcoin/Mast.java
@@ -11,11 +11,11 @@ public class Mast {
 
     public static native String generate_control_block(String jarg1, long jarg2, String jarg3);
 
-    public static String generateThresholdPubkey(String[] pubkeys, byte threshold) {
+    public static String generateThresholdPubkey(String[] pubkeys, long threshold) {
         return generate_threshold_pubkey(TextUtils.join("", pubkeys).toString(), threshold);
     }
 
-    public static String generateControlBlock(String[] pubkeys, byte threshold, String sigAggPubkey) {
+    public static String generateControlBlock(String[] pubkeys, long threshold, String sigAggPubkey) {
         return generate_control_block(TextUtils.join("", pubkeys).toString(), threshold, sigAggPubkey);
     }
 }
